+++ src/helpers/strip-object-name.ts
@@ -1,0 +1,6 @@
+import "server-only";
+
+export function stripObjectName(object: string): string | undefined {
+  const id = object.match(/(?<=:).*/g);
+  return id ? id[0] : undefined;
+}
